fix invalid config flow2
